requirements.txt
flask
flask-cors
groq
python-dotenv
colorama
twilio
spotipy
firebase-admin
certifi
